fix:rename event type file
