+++ functions/src/callable/bookings/index.ts
@@ -166,6 +166,7 @@ export const createRecurringBookingInstances = functions.https.onRequest(
           start: timestampStartDate,
           end: timestampEndDate,
           reminderDateTime: timestampReminder,
+          reminderSent: false,
           isBackgroundEvent: false,
           startDate: timestampStartDate,
           startDay: originalStartDate.format("dddd"),
@@ -243,6 +244,7 @@ export const createRecurringBookingInstances = functions.https.onRequest(
             endDate: timestampInstanceEndDate,
             endDay: instanceEndDate.format("dddd"),
             reminderDateTime: timestampInstanceReminder,
+            reminderSent: false,
             originalEventId: eventRef.id,
             isInstance: true,
             created_at: FieldValue.serverTimestamp(),
