+++ src/interfaces/event.ts
@@ -40,6 +40,7 @@ export interface EventInput {
     };
   };
   reminderDateTime?: Date;
+  reminderSent?: boolean;
   // New field: Map for tracking references of instance documents
   instanceMap?: { [instanceDate: string]: string }; // Maps instance date to document ID
   paid?: boolean; // Flag to indicate if the event is paid
